streamlit_app/requirements.txt
streamlit
numpy
pillow
tensorflow
opencv-python